sync-github-releases: move commitish resolution into the apply module
resolveTargetCommitish() decides where a to-be-created release gets tagged —
an execution concern of the apply step, not part of computing the plan. It was
only ever called by apply-release-plan.ts's resolveCreateTarget(), yet lived in
release-plan.ts.

Move it next to its sole caller and relocate its unit tests into a new
apply-release-plan.spec.ts, which also covers applyReleasePlan()'s dry-run gate
(via a fake client). release-plan.ts is now purely "compute the plan".
